requirements.txt
altgraph==0.17.3
astroid==2.15.5
colour-science==0.4.2
colour-checker-detection==0.1.4
contourpy==1.0.7
coverage==7.2.7
cycler==0.11.0
dill==0.3.6
flake8==6.0.0
fonttools==4.39.4
imageio==2.30.0
importlib-resources==5.12.0
isort==5.12.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
macholib==1.16.2
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.3
opencolorio==2.2.1
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pycodestyle==2.10.0
pyflakes==3.0.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pylint==2.17.4
pyparsing==3.0.9
pyqtgraph==0.13.3
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
requests
scipy==1.10.1
shiboken6==6.5.1.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
wrapt==1.15.0
zipp==3.15.0
